InformativeHeatmap: A Class for Enhanced Heatmaps

This class encapsulates the ComplexHeatmap package's Heatmap object, allowing for extended functionality and customizations.

Constructor for the InformativeHeatmap class. This function initializes an InformativeHeatmap object with the given data and additional parameters for heatmap customization, including visual customization based on significance levels.

# S4 method for class 'ANY'
InformativeHeatmap(
  data,
  pch_val = 16,
  unit_val = 1,
  significant_color = "black",
  trending_color = "yellow",
  significant_pvalue = 0.05,
  trending_pvalue = 0.1,
  ...
)

Arguments

data

A numeric matrix or data frame to be visualized as a heatmap.

pch_val

Plotting character (pch) value for points in the heatmap. Default is 16.

unit_val

Size of the points in the heatmap. Specified in 'mm'. Default is 1.

significant_color

Color to be used for points representing significant values. Default is "black".

trending_color

Color to be used for points representing trending values (significant but less so than those meeting the significant_pvalue criterion). Default is "yellow".

significant_pvalue

P-value threshold for significance. Points with p-values below this threshold will be colored with significant_color. Default is 0.05.

trending_pvalue

P-value threshold for trending significance. Points with p-values above significant_pvalue and below this threshold will be colored with trending_color. Default is 0.1.

...

Additional parameters passed to ComplexHeatmap::Heatmap.

Value

An object of class InformativeHeatmap with the heatmap initialized and customized according to the provided parameters.

Slots

heatmap

Heatmap object from ComplexHeatmap package.

params

List of parameters used to construct the heatmap.

Examples

# \donttest{
  data <- matrix(rnorm(100), ncol = 10)
  heatmap <- InformativeHeatmap(data, pch_val = 20, unit_val = 2,
                                significant_color = "red",
                                trending_color = "blue",
                                significant_pvalue = 0.05,
                                trending_pvalue = 0.1)
# }