Create Plot Method for ClearScatterplot

Creates a plot based on the ClearScatterplot object data.

# S4 method for class 'ClearScatterplot'
createPlot(
  object,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  negLog10pValue = 1.301,
  expressionDirection = "regulation",
  negativeLogPValue = "negLog10p",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint"
)

Arguments

object

A ClearScatterplot object.

color1

Color for one category of data points.

color2

Color for another category of data points.

color3

Color for a third category of data points.

highLog2fc

Threshold for high log2 fold change values.

lowLog2fc

Threshold for low log2 fold change values.

negLog10pValue

Threshold for -log10 p-value.

expressionDirection

Direction of gene expression.

negativeLogPValue

The name of the column containing the negative log p-values.

timeVariable

The variable representing time.

xAxis

The x-axis values.

yAxis

The y-axis values.

Value

The ClearScatterplot object with the plot updated.

Examples

plotdata <- get_clear_scatterplot_df()
scatterplotObject <- ClearScatterplot(
  data = plotdata,
  logFoldChange = "log2fc",
  timePointColumn = "timePoint",
  timePointLevels = c("T10R1", "T5R1")
)
scattered_plot <- createPlot(
  scatterplotObject,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  negLog10pValue = 1.301,
  expressionDirection = "regulation",
  negativeLogPValue = "negLog10p",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint"
)
#> Scale for colour is already present.
#> Adding another scale for colour, which will replace the existing scale.